refactor(aspect): Move support code into the C/C++ specific section
This is purely a separate commit to have a better overview of the changes.
All relevant information can be found in the previous commit moving the
aspect code.
