+++ dwyu/cc/aspect/private/preprocessing/BUILD
@@ -8,8 +8,8 @@ cc_binary(
         "//dwyu/cc/aspect/private/preprocessing/fast_parsing:process_files",
         "//dwyu/cc/aspect/private/preprocessing/wave:preprocessing_hooks",
         "//dwyu/cc/aspect/private/preprocessing/wave:process_files",
-        "//dwyu/private:program_options",
-        "//dwyu/private:utils",
+        "//dwyu/cc/private:program_options",
+        "//dwyu/cc/private:utils",
         "@boost.wave//:wave",
         "@nlohmann_json//:singleheader-json",
     ],
